synthetic_counterfactual_pairs_pipeline/requirements.txt
demucs==4.0.1
filelock==3.15.4
hydra-core==1.3.2
librosa==0.9.2
matchering==2.0.6
matplotlib==3.8.0
numpy==1.23.5
omegaconf==2.3.0
openai_whisper==20230918
parselmouth==1.1.1
pyloudnorm==0.1.1
PySoundFile==0.9.0.post1
scikit_learn==1.3.0
scipy==1.14.0
seamless_communication.egg==info
soundfile==0.12.1
torch==2.0.1
torchaudio==2.0.2
tqdm==4.66.1
transformers==4.30.2
